Restructure export type and internal types to own mod.
Make comparison of 2+ commands the default output.

Change colors of output
