+++ src/hyperfine/export/csv.rs
@@ -1,4 +1,6 @@
-use super::{ExportEntry, Exporter};
+use super::Exporter;
+
+use hyperfine::types::BenchmarkResult;
 
 use std::io::{Error, ErrorKind, Result};
 
@@ -8,7 +10,7 @@ use csv::WriterBuilder;
 pub struct CsvExporter {}
 
 impl Exporter for CsvExporter {
-    fn serialize(&self, results: &Vec<ExportEntry>) -> Result<Vec<u8>> {
+    fn serialize(&self, results: &Vec<BenchmarkResult>) -> Result<Vec<u8>> {
         let mut writer = WriterBuilder::new().from_writer(vec![]);
         for res in results {
             // The list of times can not be exported to the CSV file - remove it:
