+++ src/hyperfine/export/json.rs
@@ -1,4 +1,5 @@
-use super::{ExportEntry, Exporter};
+use super::Exporter;
+use hyperfine::types::BenchmarkResult;
 
 use std::io::{Error, ErrorKind, Result};
 
@@ -6,14 +7,14 @@ use serde_json::to_vec_pretty;
 
 #[derive(Serialize, Debug)]
 struct HyperfineSummary<'a> {
-    results: &'a Vec<ExportEntry>,
+    results: &'a Vec<BenchmarkResult>,
 }
 
 #[derive(Default)]
 pub struct JsonExporter {}
 
 impl Exporter for JsonExporter {
-    fn serialize(&self, results: &Vec<ExportEntry>) -> Result<Vec<u8>> {
+    fn serialize(&self, results: &Vec<BenchmarkResult>) -> Result<Vec<u8>> {
         let mut output = to_vec_pretty(&HyperfineSummary { results });
         for content in output.iter_mut() {
             content.push(b'\n');
