+++ src/hyperfine/mod.rs
@@ -1,8 +1,9 @@
 pub mod benchmark;
+pub mod export;
 pub mod format;
 pub mod internal;
 pub mod outlier_detection;
-pub mod warnings;
-pub mod timer;
 pub mod shell;
-pub mod export;
+pub mod timer;
+pub mod types;
+pub mod warnings;
